Move writer classes into graph-publisher
These classes are only used by the Graph Publisher and can safely be
moved into the graph-publisher plugin.
