+++ library/plugins/graph/graph-publisher/src/main/java/com/cdsap/talaiot/publisher/graph/DiskPublisher.kt
@@ -2,7 +2,7 @@ package com.cdsap.talaiot.publisher.graph
 
 import com.cdsap.talaiot.logger.LogTracker
 import com.cdsap.talaiot.publisher.Publisher
-import com.cdsap.talaiot.writer.FileWriter
+import com.cdsap.talaiot.publisher.graph.writer.FileWriter
 
 
 /**
