+++ library/plugins/graph/graph-publisher/src/main/java/com/cdsap/talaiot/publisher/graph/GexfPublisher.kt
@@ -4,7 +4,7 @@ import com.cdsap.talaiot.entities.ExecutionReport
 import com.cdsap.talaiot.entities.TaskMessageState
 import com.cdsap.talaiot.logger.LogTracker
 import com.cdsap.talaiot.publisher.graph.resources.ResourcesGexf
-import com.cdsap.talaiot.writer.FileWriter
+import com.cdsap.talaiot.publisher.graph.writer.FileWriter
 import java.util.concurrent.Executor
 
 /**
