+++ library/plugins/graph/graph-publisher/src/main/java/com/cdsap/talaiot/publisher/graph/TaskDependencyGraphPublisher.kt
@@ -3,8 +3,8 @@ package com.cdsap.talaiot.publisher.graph
 import com.cdsap.talaiot.entities.ExecutionReport
 import com.cdsap.talaiot.logger.LogTracker
 import com.cdsap.talaiot.publisher.Publisher
-import com.cdsap.talaiot.writer.DotWriter
-import com.cdsap.talaiot.writer.TaskGraphWriter
+import com.cdsap.talaiot.publisher.graph.writer.DotWriter
+import com.cdsap.talaiot.publisher.graph.writer.TaskGraphWriter
 import java.util.concurrent.Executor
 
 /**
