+++ library/plugins/graph/graph-publisher/src/test/java/com/cdsap/talaiot/publisher/graph/DotPublisherTest.kt
@@ -2,7 +2,7 @@ package com.cdsap.talaiot.publisher.graph
 
 import com.cdsap.talaiot.logger.LogTracker
 import com.cdsap.talaiot.utils.TestExecutor
-import com.cdsap.talaiot.writer.FileWriter
+import com.cdsap.talaiot.publisher.graph.writer.FileWriter
 import com.nhaarman.mockitokotlin2.*
 import io.kotlintest.specs.BehaviorSpec
