+++ library/plugins/graph/graph-publisher/src/test/java/com/cdsap/talaiot/publisher/graph/GexfPublisherTest.kt
@@ -3,7 +3,7 @@ package com.cdsap.talaiot.publisher.graph
 import com.cdsap.talaiot.publisher.graph.resources.ResourcesGexf
 import com.cdsap.talaiot.logger.LogTracker
 import com.cdsap.talaiot.utils.TestExecutor
-import com.cdsap.talaiot.writer.FileWriter
+import com.cdsap.talaiot.publisher.graph.writer.FileWriter
 import com.nhaarman.mockitokotlin2.argThat
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
