+++ library/plugins/graph/graph-publisher/src/test/java/com/cdsap/talaiot/publisher/graph/HtmlPublisherTest.kt
@@ -5,7 +5,7 @@ import com.cdsap.talaiot.logger.LogTracker
 import com.cdsap.talaiot.publisher.graph.TaskMeasurementAggregatedMock.taskMeasurementAggregated
 import com.cdsap.talaiot.publisher.graph.resources.ResourcesHtml.LEGEND_HEADER
 import com.cdsap.talaiot.utils.TestExecutor
-import com.cdsap.talaiot.writer.FileWriter
+import com.cdsap.talaiot.publisher.graph.writer.FileWriter
 import com.nhaarman.mockitokotlin2.argThat
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
